+++ packages/sdk/src/opRequestGas.ts
@@ -22,7 +22,6 @@ import { prepareOperation } from "./prepareOperation";
 import { getJoinSplitRequestTotalValue } from "./utils";
 import { SparseMerkleProver } from "./SparseMerkleProver";
 import { EthToTokenConverter } from "./conversion";
-import { Logger } from "winston";
 
 // refunds < 200k gas * gasPrice converted to gasAsset not worth refunding
 const DEFAULT_GAS_ASSET_REFUND_THRESHOLD_GAS = 200_000n;
@@ -55,7 +54,6 @@ export interface HandleOpRequestGasDeps {
   gasAssets: Map<string, Asset>;
   tokenConverter: EthToTokenConverter;
   merkle: SparseMerkleProver;
-  logger?: Logger;
 }
 
 interface GasEstimatedOperationRequest
@@ -85,9 +83,8 @@ export async function handleGasForOperationRequest(
   deps: HandleOpRequestGasDeps,
   opRequest: OperationRequest
 ): Promise<GasAccountedOperationRequest> {
-  const { logger } = deps;
   // estimate gas params for opRequest
-  logger && logger.info("estimating gas for op request");
+  console.log("estimating gas for op request");
   const { gasPrice, numJoinSplits, executionGasLimit, maxNumRefunds } =
     await estimateGasForOperationRequest(deps, opRequest);
 
@@ -116,20 +113,17 @@ export async function handleGasForOperationRequest(
       numJoinSplits * PER_JOINSPLIT_GAS +
       maxNumRefunds * PER_REFUND_GAS;
 
-    logger &&
-      logger.debug(`execution gas limit: ${executionGasLimit}`, {
-        executionGasLimit,
-      });
-    logger &&
-      logger.debug(`joinSplits gas: ${numJoinSplits * PER_JOINSPLIT_GAS}`, {
-        numJoinSplits,
-        joinSplitsGas: numJoinSplits * PER_JOINSPLIT_GAS,
-      });
-    logger &&
-      logger.debug(`refunds gas: ${maxNumRefunds * PER_REFUND_GAS}`, {
-        maxNumRefunds,
-        refundsGas: maxNumRefunds * PER_REFUND_GAS,
-      });
+    console.log(`execution gas limit: ${executionGasLimit}`, {
+      executionGasLimit,
+    });
+    console.log(`joinSplits gas: ${numJoinSplits * PER_JOINSPLIT_GAS}`, {
+      numJoinSplits,
+      joinSplitsGas: numJoinSplits * PER_JOINSPLIT_GAS,
+    });
+    console.log(`refunds gas: ${maxNumRefunds * PER_REFUND_GAS}`, {
+      maxNumRefunds,
+      refundsGas: maxNumRefunds * PER_REFUND_GAS,
+    });
 
     // attempt to update the joinSplitRequests with gas compensation
     // gasAsset will be `undefined` if the user's too broke to pay for gas
@@ -140,8 +134,7 @@ export async function handleGasForOperationRequest(
         gasEstimatedOpRequest.joinSplitRequests,
         totalGasUnitsEstimate,
         gasPrice,
-        deps.tokenConverter,
-        logger
+        deps.tokenConverter
       );
 
     if (!gasAssetAndTicker) {
@@ -178,8 +171,7 @@ async function tryUpdateJoinSplitRequestsForGasEstimate(
   joinSplitRequests: JoinSplitRequest[],
   gasUnitsEstimate: bigint,
   gasPrice: bigint,
-  tokenConverter: EthToTokenConverter,
-  logger?: Logger
+  tokenConverter: EthToTokenConverter
 ): Promise<[JoinSplitRequest[], AssetAndTicker | undefined]> {
   // group joinSplitRequests by asset address
   const joinSplitRequestsByAsset = groupByMap(
@@ -226,11 +218,10 @@ async function tryUpdateJoinSplitRequestsForGasEstimate(
         matchingJoinSplitRequests
       );
 
-      logger &&
-        logger.debug(
-          `amount to add for gas asset by modifying existing joinsplit for gas asset with ticker ${ticker}: ${estimateInGasAsset}}`,
-          { gasAssetTicker: ticker, gasAsset, estimateInGasAsset }
-        );
+      console.log(
+        `amount to add for gas asset by modifying existing joinsplit for gas asset with ticker ${ticker}: ${estimateInGasAsset}}`,
+        { gasAssetTicker: ticker, gasAsset, estimateInGasAsset }
+      );
 
       return [
         Array.from(joinSplitRequestsByAsset.values()).flat(),
@@ -259,15 +250,14 @@ async function tryUpdateJoinSplitRequestsForGasEstimate(
         unwrapValue: estimateInGasAssetIncludingNewJoinSplit,
       });
 
-      logger &&
-        logger.debug(
-          `amount to add for gas asset by adding a new joinsplit for gas asset with ticker ${ticker}: ${estimateInGasAssetIncludingNewJoinSplit}`,
-          {
-            gasAssetTicker: ticker,
-            gasAsset,
-            estimateInGasAssetIncludingNewJoinSplit,
-          }
-        );
+      console.log(
+        `amount to add for gas asset by adding a new joinsplit for gas asset with ticker ${ticker}: ${estimateInGasAssetIncludingNewJoinSplit}`,
+        {
+          gasAssetTicker: ticker,
+          gasAsset,
+          estimateInGasAssetIncludingNewJoinSplit,
+        }
+      );
 
       return [modifiedJoinSplitRequests, { asset: gasAsset, ticker }];
     }
@@ -279,7 +269,7 @@ async function tryUpdateJoinSplitRequestsForGasEstimate(
 
 // estimate gas params for opRequest
 async function estimateGasForOperationRequest(
-  { handlerContract, logger, ...deps }: HandleOpRequestGasDeps,
+  { handlerContract, ...deps }: HandleOpRequestGasDeps,
   opRequest: OperationRequest
 ): Promise<GasParams> {
   let { executionGasLimit, maxNumRefunds, gasPrice } = opRequest;
@@ -310,7 +300,7 @@ async function estimateGasForOperationRequest(
 
   // simulate the operation
   if (!executionGasLimit || !maxNumRefunds) {
-    logger && logger.info("simulating operation");
+    console.log("simulating operation");
     const result = await simulateOperation(
       handlerContract,
       preparedOp as PreSignOperation